requirement.txt
matplotlib
opencv-python
opencv-contrib-python
pandas
pyautogui
Pillow
requests
firebase_admin